Rename time-series beast package to runtime
