+++ src/dr/inferencexml/timeseries/SyntheticOUTimeSeriesBenchmarkParser.java
@@ -6,9 +6,9 @@ import dr.inference.model.GivensRotationMatrixParameter;
 import dr.inference.model.MatrixParameter;
 import dr.inference.model.OrthogonalBlockDiagonalPolarStableMatrixParameter;
 import dr.inference.model.Parameter;
-import dr.inference.timeseries.beast.CompositeGradientWrtParameterProvider;
-import dr.inference.timeseries.beast.TimeSeriesBenchmark;
-import dr.inference.timeseries.beast.TimeSeriesGradient;
+import dr.inference.timeseries.runtime.CompositeGradientWrtParameterProvider;
+import dr.inference.timeseries.runtime.TimeSeriesBenchmark;
+import dr.inference.timeseries.runtime.TimeSeriesGradient;
 import dr.inference.timeseries.core.BasicTimeSeriesModel;
 import dr.inference.timeseries.core.IrregularTimeGrid;
 import dr.inference.timeseries.core.TimeGrid;
